+++ src/background/modules/system_settings/mod.rs
@@ -2,3 +2,4 @@ pub mod application;
 mod domain;
 pub mod infrastructure;
 pub mod language;
+pub(crate) mod nightlight;