+++ changelog.md
@@ -6,6 +6,7 @@
 
 - HDR toggle for supported monitors on quick settings widget.
 - Light/dark mode toggle on quick settings widget.
+- Night Light toggle on quick settings widget.
 
 ### fix